extract funnel metric logic
